+++ location.h
@@ -12,16 +12,16 @@ class Location
     QString city() { return m_city; }
     QString country() { return m_country; }
     QString street() { return m_street; }
-    int latitude() { return m_lat; }
-    int longitude() { return m_long; }
+    qreal latitude() { return m_lat; }
+    qreal longitude() { return m_long; }
 
     private:
       QString m_name;
       QString m_city;
       QString m_country;
       QString m_street;
-      int m_lat;
-      int m_long;
+      qreal m_lat;
+      qreal m_long;
 };
 
 #endif // LOCATION_H
